Initial (and only) commit
A push of the app I made for my work to aid in inventorying parts.  It
had some UI issues with the car information page, and the table had
trouble updating switch states, but it generally performed wonderfully.
 Live interaction with the SQL server coupled with a clean interface
made it a great tool to use, but unfortunately it was made for the
wrong sort of people at the wrong time.